requirements.txt
flask~=3.0.2
requests~=2.31.0
python-telegram-bot
langchain
openai
python-dotenv~=1.0.1
icecream~=2.1.3
langchain-openai
langchain-pinecone
langchain-community
ctransformers